+++ src/CHANGELOG.md
@@ -277,7 +277,7 @@ Features:
 - Added list view for albums, artists, playlists, and collections.
 - Tweaked sorting options and styling for compatability with list views.
 - Better handling of various artists.
-- Display "appears on" compilation albums in artist view, where possible.
+- Display "appears on" albums in artist view, where possible.
 
 <!-- CHANGELOG SPLIT MARKER -->